Inline eth_simulateV1 transfer tracer into the keeper
The tracer used by eth_simulateV1 to capture EVM logs and synthesize
ERC-7528 native-Transfer events lived in its own x/evm/tracer/transfertracer
package, but had only one consumer (the keeper's simulate driver) and was
never registered with go-ethereum's pluggable tracer directory. Move it
into the keeper package as simulate_v1_logtracer.go, modeled on upstream
go-ethereum's internal/ethapi/logtracer.go which co-locates its analogous
helper next to simulate.go.

Drop the package-local denyList in favour of the existing
mezoCustomPrecompileAddrs set in state_override.go, since both build the
same map from types.DefaultPrecompilesVersions.
